requirements.txt
opencv-python
mediapipe
scikit-learn
pyserial
numpy
flask
flask-socketio
eventlet
